backend-prompt-toolkit/requirements.txt
alembic==1.14.1
aniso8601==10.0.1
asgiref==3.8.1
babel==2.17.0
bcrypt==4.3.0
better-profanity==0.7.0
blinker==1.8.2
capstone==6.0.0a5
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colored-traceback==0.4.2
cryptography==46.0.3
distlib==0.3.9
ecdsa==0.19.1
enum-compat==0.0.3
et_xmlfile==2.0.0
Faker==35.2.2
filelock==3.16.1
Flask==3.0.3
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-Builder==0.9
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Redmail==0.3.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
intervaltree==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
libscca-python==20221027
lznt1==0.2
mailgun==0.1.1
Mako==1.3.10
MarkupSafe==2.1.5
mpmath==1.3.0
numpy==1.24.4
openpyxl==3.1.5
packaging==25.0
paramiko==3.5.1
pipenv==2024.4.1
platformdirs==4.3.6
plumbum==1.9.0
prefetch-parser==1.0.0
profanity-check==1.0.3
psutil==7.1.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pwntools==4.15.0
pyactiveresource==1.0.1
pycparser==2.23
pyelftools==0.32
Pygments==2.19.2
PyJWT==2.9.0
PyMuPDF==1.24.11
PyNaCl==1.6.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-registry==1.3.1
pytz==2025.2
raven==6.10.0
redmail==0.6.0
requests==2.32.4
ROPGadget==7.7
rpyc==6.0.2
scikit-learn==1.3.2
scipy==1.10.1
sendgrid==6.12.4
six==1.17.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
SQLAlchemy-serializer==1.4.12
sympy==1.13.3
threadpoolctl==3.5.0
typing_extensions==4.13.2
unicodecsv==0.14.1
unicorn==2.1.2
unix-ar==0.2.1
urllib3==2.2.3
virtualenv==20.31.2
Werkzeug==3.0.6
zipp==3.20.2
zstandard==0.23.0
